pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "orjson>=3.10.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.0",
  "redis>=5.0.0",
  "SQLAlchemy>=2.0.30",
  "asyncmy>=0.2.9",
  "rocketmq-client-python==2.0.0",
  "grpcio>=1.62.0",
  "grpcio-tools>=1.62.0",
  "langchain>=0.2.0",
  "langchain-core>=0.2.0",
  "langchain-openai>=1.1.6",
  "tiktoken<0.12",
]

[project.scripts]
chatbot-bot = "chatbot.main:main"
chatbot-scheduler = "chatbot.workers.send_scheduler_worker:main"

[tool.uv]
package = true
